requirements.txt
torch>=2.0.0
numpy>=1.21.0
pysam>=0.19.0
biopython>=1.79
tqdm>=4.62.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
scikit-learn>=1.0.0
pathlib>=1.0.0




